add withdraw

+++ deploy.sh
@@ -1,6 +1,13 @@
 rm -f ./build/contracts/Auction.json ./build/contracts/TreeCoin.json 
-truffle deploy
-rm -f ./app/src/Auction.json
-rm -f ./app/src/TreeCoin.json
-cp -f ./build/contracts/Auction.json ./app/src
-cp -f ./build/contracts/TreeCoin.json ./app/src
+
+a_var= truffle deploy | grep "Error"
+if  [ "$a_var" = "\n" ]
+then
+    echo $a_var
+    echo "error"
+else
+    rm -f ./app/src/Auction.json
+    rm -f ./app/src/TreeCoin.json
+    cp -f ./build/contracts/Auction.json ./app/src
+    cp -f ./build/contracts/TreeCoin.json ./app/src
+fi
